refactor WindVectors
